refactor useSearch hook for better performance and maintainability
